move clusterQuery helper function
