+++ api/cluster.go
@@ -1,10 +1,12 @@
 package api
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"net/http"
 	"strconv"
+	"sync"
 
 	"github.com/grafana/metrictank/api/middleware"
 	"github.com/grafana/metrictank/api/models"
@@ -183,4 +185,70 @@ func (s *Server) indexDelete(ctx *middleware.Context, req models.IndexDelete) {
 		DeletedDefs: len(defs),
 	}
 	response.Write(ctx, response.NewMsgp(200, &resp))
+}
+
+// clusterQuery takes a request and the path to request it on, then fans it out
+// across the cluster, except to the local peer.
+// ctx:          request context
+// data:         request to be submitted
+// name:         name to be used in logging & tracing
+// path:         path to request on
+// respTemplate: the response object into which the responses shall
+//               be deserialized. each sub-request will get a copy of this
+//               template and the response will be unarshalled into that copy.
+//               (generic without generics)
+func (s *Server) clusterQuery(ctx context.Context, data cluster.Traceable, name, path string, respTemplate msgp.Unmarshaler) ([]msgp.Unmarshaler, error) {
+	peers, err := cluster.MembersForQuery()
+	if err != nil {
+		log.Error(3, "HTTP clusterQuery unable to get peers, %s", err)
+		return nil, err
+	}
+	log.Debug("HTTP %s across %d instances", name, len(peers)-1)
+
+	result := make([]msgp.Unmarshaler, 0, len(peers)-1)
+
+	var errors []error
+	var errLock sync.Mutex
+	var resLock sync.Mutex
+	var wg sync.WaitGroup
+	for _, peer := range peers {
+		if peer.IsLocal() {
+			continue
+		}
+		wg.Add(1)
+		go func(peer cluster.Node) {
+			defer wg.Done()
+			log.Debug("HTTP Render querying %s%s", peer.Name, path)
+			buf, err := peer.Post(ctx, name, path, data)
+
+			if err != nil {
+				log.Error(4, "HTTP Render error querying %s%s: %q", peer.Name, path, err)
+				errLock.Lock()
+				errors = append(errors, err)
+				errLock.Unlock()
+				return
+			}
+
+			resp := respTemplate
+			_, err = resp.UnmarshalMsg(buf)
+			if err != nil {
+				log.Error(4, "HTTP error unmarshaling body from %s%s: %q", peer.Name, path, err)
+				errLock.Lock()
+				errors = append(errors, err)
+				errLock.Unlock()
+				return
+			}
+
+			resLock.Lock()
+			result = append(result, resp)
+			resLock.Unlock()
+		}(peer)
+	}
+	wg.Wait()
+
+	if len(errors) > 0 {
+		return nil, errors[0]
+	}
+
+	return result, nil
 }
